refactor(all): move reusable/shared analytics types/interfaces into @lace/common
