+++ packages/common/src/index.ts
@@ -28,3 +28,5 @@ export * from '@ui/components/Switch';
 export * from '@ui/components/Loader';
 export * from '@ui/components/Radio';
 export * from '@ui/components/Timeline';
+
+export * from '@analytics/index';
