+++ packages/staking/src/features/browse-pools/StakePoolsTable/StakePoolTableBrowser/StakePoolTableBrowser.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable react/no-multi-comp */
 import Icon from '@ant-design/icons';
 import { Wallet } from '@lace/cardano';
+import { PostHogAction } from '@lace/common';
 import { List, ListProps, Tooltip } from 'antd';
 import cn from 'classnames';
 import isNumber from 'lodash/isNumber';
@@ -8,7 +9,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import InfiniteScroll from 'react-infinite-scroll-component';
 import InfoIcon from '../../../../assets/icons/info-icon.svg';
-import { PostHogAction, useOutsideHandles } from '../../../outside-handles-provider';
+import { useOutsideHandles } from '../../../outside-handles-provider';
 import { useDelegationPortfolioStore } from '../../../store';
 import { StakePoolItemBrowser, StakePoolItemBrowserProps } from '../StakePoolItemBrowser';
 import Arrow from './arrow.svg';
