+++ packages/staking/src/features/drawer/preferences/StepPreferencesFooter.tsx
@@ -1,6 +1,7 @@
+import { PostHogAction } from '@lace/common';
 import { Button, Flex } from '@lace/ui';
 import { Tooltip } from 'antd';
-import { PostHogAction, useOutsideHandles } from 'features/outside-handles-provider';
+import { useOutsideHandles } from 'features/outside-handles-provider';
 import { useDelegationPortfolioStore } from '../../store';
 
 type StepPreferencesFooterProps = {
